+++ accounting/src/sdk/models/shared/createjournalresponse.ts
@@ -14,6 +14,9 @@ import { Expose, Type } from "class-transformer";
  * Success
  */
 export class CreateJournalResponse extends SpeakeasyBase {
+    /**
+     * Contains a single entry that communicates which record has changed and the manner in which it changed.
+     */
     @SpeakeasyMetadata({ elemType: PushOperationChange })
     @Expose({ name: "changes" })
     @Type(() => PushOperationChange)
@@ -153,7 +156,7 @@ export class CreateJournalResponse extends SpeakeasyBase {
     timeoutInMinutes?: number;
 
     /**
-     * @deprecated this field will be removed in a future release, please migrate away from it as soon as possible
+     * @deprecated field: This will be removed in a future release, please migrate away from it as soon as possible.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "timeoutInSeconds" })