+++ accounting/src/sdk/models/shared/directcostlineitem.ts
@@ -104,7 +104,7 @@ export class DirectCostLineItem extends SpeakeasyBase {
     /**
      * Collection of categories against which this direct cost is tracked.
      *
-     * @deprecated this field will be removed in a future release, please migrate away from it as soon as possible
+     * @deprecated field: This will be removed in a future release, please migrate away from it as soon as possible.
      */
     @SpeakeasyMetadata({ elemType: TrackingCategoryRef })
     @Expose({ name: "trackingCategoryRefs" })