+++ accounting/src/sdk/models/shared/invoicelineitem.ts
@@ -107,7 +107,7 @@ export class InvoiceLineItem extends SpeakeasyBase {
     /**
      * Reference to the tracking categories to which the line item is linked.
      *
-     * @deprecated this field will be removed in a future release, please migrate away from it as soon as possible
+     * @deprecated field: This will be removed in a future release, please migrate away from it as soon as possible.
      */
     @SpeakeasyMetadata({ elemType: TrackingCategoryRef })
     @Expose({ name: "trackingCategoryRefs" })